+++ src/main/java/com/knu/KnowcKKnowcK/authenticationHandler/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.knu.KnowcKKnowcK.exception.jwtHandler;
+package com.knu.KnowcKKnowcK.authenticationHandler;
 
 
 import jakarta.servlet.ServletException;
